Fix flakiness in `test_stop_workers_politely`
The `test_stop_workers_politely` test has been a bit flaky. Given the
test is trying to shutdown a worker and compute results at the same
time, this seems like part of the problem. To try and fix this, wait
until the worker is shutdown completely before trying to retrieve
results. Otherwise it seems to have issues.