Cleanup after Miniconda install in Docker build
To cutdown on the Docker image size, handle the install of Miniconda and
subsequent clean up in the same `RUN` step. The cleanup includes the
removal the install script, cached conda packages, and conda package
tarballs. This should cutdown the size of this Docker layer and save
memory on CI.