Update `scale_down` to shutdown workers with DRMAA
As the purpose of `scale_down` is to shutdown workers through the
scheduler directly, pull code from `stop_workers` to do exactly this
with DRMAA. Update `stop_workers` to make use of `scale_down` for this
functionality as well. Should help ensure `DRMAACluster` matches more
closely to the expected `Cluster` specification.