Call `dask-worker` through Python executable (#91)
To ensure that the right `dask-worker` executable is found, make use of
`python -m` to search for and run the `distributed.cli.dask_worker`
module. This is robust to different install locations or directory
layouts as it doesn't care where `dask-worker` actually is. Instead it
only cares where `distributed` is installed, which has to be accessible
from the same `python` running `dask-drmaa` (as it is a dependency).
From there it can simply rely on `python` to figure out where
`dask_worker` lives so that it can run it.