Mark `test_stop_single_worker` as a known failure (#94)
This test has been flaky historically. However it appears to now be
reliably failing on both Python 2 & 3. Would be good to fix it once the
cause is understood. Though it is not a critical failure, so we can
tolerate marking it as a failure and tracking for now. There are other
changes planned (e.g. updating worker startup and shutdown) that may end
up fixing it.